backend/requirements.txt
django==3.*
djangorestframework==3.12.*
python-dotenv==0.19.*
requests==2.28.*
django-crontab==0.5.*
django-cors-headers==4.4.*
drf-yasg==1.21.*
coreapi==2.3.*